requirements.txt
pytest
jsonpickle
pony
psycopg2
requests
allure-pytest
jsonschema



















